Factorize lolp vcard
