+++ src/solver/variable/include/antares/solver/variable/economy/lolp_base.h
@@ -1,0 +1,266 @@
+/*
+ * Copyright 2007-2025, RTE (https://www.rte-france.com)
+ * See AUTHORS.txt
+ * SPDX-License-Identifier: MPL-2.0
+ * This file is part of Antares-Simulator,
+ * Adequacy and Performance assessment for interconnected energy networks.
+ *
+ * Antares_Simulator is free software: you can redistribute it and/or modify
+ * it under the terms of the Mozilla Public Licence 2.0 as published by
+ * the Mozilla Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * Antares_Simulator is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Mozilla Public Licence 2.0 for more details.
+ *
+ * You should have received a copy of the Mozilla Public Licence 2.0
+ * along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
+ */
+
+#pragma once
+
+#include <antares/memory/memory.h>
+#include <antares/solver/variable/categories.h>
+#include <antares/solver/variable/container.h>
+#include <antares/solver/variable/state.h>
+#include <antares/solver/variable/storage/intermediate.h>
+#include <antares/solver/variable/storage/results.h>
+#include <antares/solver/variable/surveyresults.h>
+#include <antares/solver/variable/variable.h>
+#include <antares/study/area/area.h>
+#include <antares/study/area/links.h>
+#include <antares/study/study.h>
+
+namespace Antares::Solver::Variable::Economy
+{
+
+template<class Traits>
+struct VCardLOLP_Base
+{
+    //! Caption
+    static std::string Caption()
+    {
+        return Traits::Caption();
+    }
+
+    //! Unit
+    static std::string Unit()
+    {
+        return Traits::Unit();
+    }
+
+    //! The short description of the variable
+    static std::string Description()
+    {
+        return Traits::Description();
+    }
+
+    //! The expecte results
+    typedef Results<R::AllYears::Average< // The average values throughout all years
+      >>
+      ResultsType;
+
+    typedef VCardLOLP_Base VCardForSpatialAggregate;
+
+    static constexpr uint8_t categoryDataLevel = Category::DataLevel::area;
+    //! File level (provided by the type of the results)
+    static constexpr uint8_t categoryFileLevel = ResultsType::categoryFile
+                                                 & (Category::FileLevel::id
+                                                    | Category::FileLevel::va);
+    //! Precision (views)
+    static constexpr uint8_t precision = Category::all;
+    //! Indentation (GUI)
+    static constexpr uint8_t nodeDepthForGUI = +0;
+    //! Decimal precision
+    static constexpr uint8_t decimal = 2;
+    //! Number of columns used by the variable (One ResultsType per column)
+    static constexpr int columnCount = 1;
+    //! The Spatial aggregation
+    static constexpr uint8_t spatialAggregate = Traits::spatialAggregate;
+    static constexpr uint8_t spatialAggregateMode = Category::spatialAggregateEachYear;
+    static constexpr uint8_t spatialAggregatePostProcessing = 0;
+    //! Intermediate values
+    static constexpr uint8_t hasIntermediateValues = 1;
+    //! Can this variable be non applicable (0 : no, 1 : yes)
+    static constexpr uint8_t isPossiblyNonApplicable = 0;
+
+    typedef IntermediateValues IntermediateValuesBaseType;
+    typedef std::vector<IntermediateValues> IntermediateValuesType;
+
+    using IntermediateValuesTypeForSpatialAg = std::unique_ptr<IntermediateValuesBaseType[]>;
+
+}; // class VCard
+
+/*!
+** \brief Base class for LOLP and LOLP_CSR variables
+*/
+template<class Traits, class NextT = Container::EndOfList>
+class LOLP_Base: public Variable::IVariable<LOLP_Base<Traits, NextT>, NextT, VCardLOLP_Base<Traits>>
+{
+public:
+    //! Type of the next static variable
+    typedef NextT NextType;
+    //! VCard
+    typedef VCardLOLP_Base<Traits> VCardType;
+    //! Ancestor
+    typedef Variable::IVariable<LOLP_Base<Traits, NextT>, NextT, VCardType> AncestorType;
+
+    //! List of expected results
+    typedef typename VCardType::ResultsType ResultsType;
+
+    typedef VariableAccessor<ResultsType, VCardType::columnCount> VariableAccessorType;
+
+    enum
+    {
+        //! How many items have we got
+        count = 1 + NextT::count,
+    };
+
+    template<int CDataLevel, int CFile>
+    struct Statistics
+    {
+        enum
+        {
+            count = ((VCardType::categoryDataLevel & CDataLevel
+                      && VCardType::categoryFileLevel & CFile)
+                       ? (NextType::template Statistics<CDataLevel, CFile>::count
+                          + VCardType::columnCount * ResultsType::count)
+                       : NextType::template Statistics<CDataLevel, CFile>::count),
+        };
+    };
+
+public:
+    void initializeFromStudy(Data::Study& study)
+    {
+        pNbYearsParallel = study.maxNbYearsInParallel;
+
+        // Intermediate values
+        InitializeResultsFromStudy(AncestorType::pResults, study);
+
+        pValuesForTheCurrentYear.resize(pNbYearsParallel);
+        for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
+        {
+            pValuesForTheCurrentYear[numSpace].initializeFromStudy(study);
+        }
+        // Next
+        NextType::initializeFromStudy(study);
+    }
+
+    template<class R>
+    static void InitializeResultsFromStudy(R& results, Data::Study& study)
+    {
+        VariableAccessorType::InitializeAndReset(results, study);
+    }
+
+    void initializeFromArea(Data::Study* study, Data::Area* area)
+    {
+        // Next
+        NextType::initializeFromArea(study, area);
+    }
+
+    void initializeFromLink(Data::Study* study, Data::AreaLink* link)
+    {
+        // Next
+        NextType::initializeFromAreaLink(study, link);
+    }
+
+    void simulationBegin()
+    {
+        for (unsigned int numSpace = 0; numSpace < pNbYearsParallel; numSpace++)
+        {
+            pValuesForTheCurrentYear[numSpace].reset();
+        }
+        // Next
+        NextType::simulationBegin();
+    }
+
+    void simulationEnd()
+    {
+        NextType::simulationEnd();
+    }
+
+    void yearBegin(unsigned int year, unsigned int numSpace)
+    {
+        // Reset the values for the current year
+        pValuesForTheCurrentYear[numSpace].reset();
+
+        // Next variable
+        NextType::yearBegin(year, numSpace);
+    }
+
+    void yearEndBuild(State& state, unsigned int year, unsigned int numSpace)
+    {
+        // Next variable
+        NextType::yearEndBuild(state, year, numSpace);
+    }
+
+    void yearEnd(unsigned int year, unsigned int numSpace)
+    {
+        // Compute all statistics for the current year (daily,weekly,monthly)
+        pValuesForTheCurrentYear[numSpace].computeStatisticsOrForTheCurrentYear();
+
+        // Next variable
+        NextType::yearEnd(year, numSpace);
+    }
+
+    void computeSummary(unsigned int year, unsigned int numSpace)
+    {
+        // Merge all those values with the global results
+        AncestorType::pResults.merge(year, pValuesForTheCurrentYear[numSpace]);
+
+        // Next variable
+        NextType::computeSummary(year, numSpace);
+    }
+
+    void hourBegin(unsigned int hourInTheYear)
+    {
+        // Next variable
+        NextType::hourBegin(hourInTheYear);
+    }
+
+    void hourForEachArea(State& state, unsigned int numSpace)
+    {
+        if (Traits::checkCondition(state))
+        {
+            pValuesForTheCurrentYear[numSpace][state.hourInTheYear] = 100;
+        }
+
+        // Next variable
+        NextType::hourForEachArea(state, numSpace);
+    }
+
+    Antares::Memory::Stored<double>::ConstReturnType retrieveRawHourlyValuesForCurrentYear(
+      unsigned int,
+      unsigned int numSpace) const
+    {
+        return pValuesForTheCurrentYear[numSpace].hour;
+    }
+
+    void localBuildAnnualSurveyReport(SurveyResults& results,
+                                      int fileLevel,
+                                      int precision,
+                                      unsigned int numSpace) const
+    {
+        // Initializing external pointer on current variable non applicable status
+        results.isCurrentVarNA = AncestorType::isNonApplicable;
+
+        if (AncestorType::isPrinted[0])
+        {
+            // Write the data for the current year
+            results.variableCaption = VCardType::Caption();
+            results.variableUnit = VCardType::Unit();
+            pValuesForTheCurrentYear[numSpace]
+              .template buildAnnualSurveyReport<VCardType>(results, fileLevel, precision);
+        }
+    }
+
+private:
+    //! Intermediate values for each year
+    typename VCardType::IntermediateValuesType pValuesForTheCurrentYear;
+    unsigned int pNbYearsParallel;
+
+}; // class LOLP_Base
+
+} // namespace Antares::Solver::Variable::Economy
